added test and convolutions
